requirements.txt
certifi==2021.5.30
chardet==4.0.0
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
greenlet==3.0.3
grpcio==1.63.0
grpcio-tools==1.63.0
hupper==1.12.1
idna==2.10
Jinja2==3.0.1
kiwisolver==1.4.5
MarkupSafe==2.0.1
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pillow==10.3.0
protobuf==5.26.1
pyparsing==3.1.2
pyserial==3.5
python-dateutil==2.9.0.post0
requests==2.25.1
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
urllib3==1.26.6
